Corrected sizeWithFont calculation error caused by unnecessary end-of-line spaces
Added a condition for the first word in each line to not add a space,
and then subsequent words add the space before.  This makes it so that
the last word on each line doesn't have an ending space in its size
calculation.  I was running into issues where the number of lines
calculations were not accurate due to this space.